Fix 64bit detection, broken since 66e69c4c25ef2554b3ed44378dc5fc465ab47781. (#34)
This commit restores the previous logic of accepting an 8-byte intptr as evidence
enough for 64bit. It also accepts presence of IsWow64Process as evidence of 64bit.
On 64bit, this may return true or false depending on the process. On 32bit this entry
point is not available and an exception will be raised.