Try to avoid a crash
Sometimes the activated FTD2XX component causes a serious
crash in the runtime, due to a buffer overflow.
Using a different method seems to avoid the problem.
Even though the library doesn't seem to write past
the required buffer size, it still writes more than necessary.
Maybe the function with the structure is even worse.