Ignore temperature updates if the sensor is NONE
This commit adds a condition to skip updating the sensor value if the
other fields are not valid.

For whatever reason the temperature is coming back as 0 at times.  When
this happens the controller is also reported as NONE, the min==max==0,
and the Target is also NONE.  (Only the first condition is checked)